cloud_functions/delete_inter_files/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage>=1.30.0
google-cloud-bigquery>=1.26.1